chore: curl_utils
