+++ engine/cli/CMakeLists.txt
@@ -85,7 +85,8 @@ add_executable(${TARGET_NAME} main.cc
     ${CMAKE_CURRENT_SOURCE_DIR}/utils/easywsclient.cc
     ${CMAKE_CURRENT_SOURCE_DIR}/utils/download_progress.cc
     ${CMAKE_CURRENT_SOURCE_DIR}/../utils/config_yaml_utils.cc
-    ${CMAKE_CURRENT_SOURCE_DIR}/../utils/file_manager_utils.cc 
+    ${CMAKE_CURRENT_SOURCE_DIR}/../utils/file_manager_utils.cc
+    ${CMAKE_CURRENT_SOURCE_DIR}/../utils/curl_utils.cc 
   )
 
 target_link_libraries(${TARGET_NAME} PRIVATE CLI11::CLI11)
