+++ engine/cli/commands/engine_list_cmd.cc
@@ -4,6 +4,7 @@
 #include "common/engine_servicei.h"
 #include "server_start_cmd.h"
 #include "utils/curl_utils.h"
+#include "utils/engine_constants.h"
 #include "utils/logging_utils.h"
 #include "utils/url_parser.h"
 // clang-format off
