+++ engine/cli/commands/server_start_cmd.cc
@@ -1,6 +1,7 @@
 #include "server_start_cmd.h"
 #include "commands/cortex_upd_cmd.h"
 #include "utils/cortex_utils.h"
+#include "utils/engine_constants.h"
 #include "utils/file_manager_utils.h"
 #include "utils/widechar_conv.h"
